Add Announcements instance on player creation and clean-up PlayerDto fields
A new instance of Announcements is now created during the player registration process to ensure each player has a fresh set of announcements. This modification has been applied both to normal and external registrations. In addition, 'EloStateBeforeMatch' field was removed from PlayerDto, as it was deprecated and no longer necessary in the application logic. Moreover, the default initialization was added to the 'Announcements' property in the 'Player' model to prevent potential null reference exceptions.