Remove redundant CSS link and enhance user authentication
This commit removes a redundant CSS link in the index.html file that was needlessly increasing page load
time. Changes have also been made in MainLayout.razor to improve user authentication by checking if the
user is authenticated and fetching necessary information during initialization. This is done by injecting
AuthenticationStateProvider, IToastService, and HttpClient dependencies and using them in the OnInitializedAsync
method.

In addition, a new file named AnnouncementController.cs, which checks whether the initial level is done or not,
is added to improve the gaming progress tracking. Lastly, adjustments are made to Notifications.razor styling,
and a minor clean-up was done in NavMenu.razor.