Add country to user's information
The country property has been added to the GetAllUsersQueryHandler class to include country data in the user information. This was done to enrich the user profile details on the UsersList component. In addition to that, the front-end has also been updated to display the user's country in the user list.