1Implement user registration date setting and refactor Startup
A user registration date setting flow is introduced in the ExternalLogin.cshtml.cs file, providing a timestamp for when a user is registered. Also, significant refactoring and error checking is done in the Startup.cs file. The ConfigureServices method is made static, UI cookie configuration is modified, and stricter checks are introduced for Google OAuth Settings. Refactoring also goes ahead to change how Startup class methods are utilized in the Program.cs file.