Add ReturnUrl support for login redirect
A ReturnUrl parameter has been added to the MatchDetailPlayers component, which is used to redirect users to the login page when they are not authenticated. The ReturnUrl gets its value from the current match page URL, specifically the MatchId and detail. This change ensures that after a successful login, the user will be redirected back to the page they were trying to access.