Remove toast error message when registering for a match
This commit removes the redundant error toast message "You must be logged in to register for this match". This decision was made to streamline the user registration flow. Now, when a user attempts to register for a match without being authenticated, the user is directly navigated to the login page, therefore the toast error message became unnecessary.