"Updated UI and logic to improve user experience"
This commit includes a series of changes designed to improve user experience. The changes were applied to the "InitialLevelForm" and "Index" components. In the "InitialLevelForm", the texts were tweaked and some steps in the form were renamed for clarity. Furthermore, I improved the modality operation logic so it won't open if the initial questionnaire is already filled. For the "Index" component, the InitialLevelForm and DevelopmentAnnouncement are no longer always on display; they now only appear when they have not been completed/read by the player. These changes improve system efficiency and usability, making for a smoother user experience.