"Refactored check for unread announcements and initial level form
The announcement checks were moved from the MainLayout to the Index page, and now use separate classes, InitialLevelResponse and AnnouncementDevelopmentResponse, to store the state of the initial level form and development announcement respectively, rather than the PlayerDto class. This change was made to ensure that the information about the user is acquired from the server only when it's truly needed - when we're trying to render the Index page. It reduces unnecessary loads when accessing other parts of the application, thereby improving performance.

Also, two new routes 'initial-level-done' and 'development-announcement-read-it' have been added to the AnnouncementController in the server, returning