Add ability to delete calibrations and show registration dates
A new method, DeleteAllInitialPlayerCalibration, has been added in PlayerController which allows administrators to delete all initial calibration data for players. A button to trigger this action is now available on Users Manager page, with confirm dialog and toast notifications for success/error. RegistrationDate field was added in UserDto and employed in GetAllUsersQueryHandler for sorting returned users by this date descending. In UsersList component on frontend, email column has been replaced with registration date.