Add Announcements to PlayerController fetch query
Modified the query in PlayerController used to fetch player data to also include any announcements associated with that player. This change was necessary to ensure all relevant player data is retrieved in one single operation, reducing the need for additional queries and improving overall performance. Furthermore, it helps in case of initial level form completion, to set the announcement flag to false.