Adjust initial ELO point allocation for player skill levels
Updated the ELO points rewarded in InitialPlayerSkillCalibrationCommandHandler.cs for racket sports and Padel skill levels to better balance the game. For "SemiPro" and "Pro" racket sports levels, points reduced from 100 to 75 and 200 to 150 respectively. In contrast, for Padel skill levels, including "BeginnerIntermediate" to "Pro", the points were adjusted with a specific formula (90 + (350 - 90) / 2^n) to ensure a fairer distribution of ELO points. The "Pro" level points were capped at 350. This change was made to promote fair competition and ensure player progression.