Add share icon to match details and update initial player skill calculation
In the MatchDetail.razor file, a share icon was added to allow users to more easily share match details. On the server side within the InitialPlayerSkillCalibrationCommandHandler, the method of calculating a new player’s initial skill was adjusted. Previously, players with past experience in other racket sports were given a lump sum of 50 extra Elo points. Now, instead of a lump sum, points are awarded based on the number of years spent playing the other sport, up to a maximum of five years. This was done to better balance initial player skill values.