Add share functionality to MatchDetail
A ClipboardService has been added to handle the copying of event URLs for sharing purposes. The match detail page now includes a share button that uses this ClipboardService to copy the current URL to the user's clipboard so that the user can easily share the event. Modification in the layout was required to accommodate this new functionality. Also, a new service is scoped in the project file.