Add initial level check to main layout
Expanded main layout to include an initial level check. This change allows the application to consistently monitor if the initial level has been done for the user across all pages, instead of just on the index page where this function was previously located. The change allows the application to react more promptly when the user completes the initial level, improving the user experience.