Add Matches Manager and remove calibration deletion in bulk
This commit includes the addition of a Matches Manager, which will help keep track of all matches in the system. This new manager can be accessed through the navigation menu. Also, the ability to delete all player calibrations at once was removed, to prevent any unintended massive data loss. Now, calibrations must be deleted individually within each user's details. Other clean-up and refinements were made to the existing code base as a part of this commit.