"Add administrator match management functionality
A new API endpoint and view were added to allow administrators to access all match data and perform actions like deleting matches. An authorization for the role of 'Administrator' was added to the new API endpoint in MatchController. The view in MatchesManager.razor was updated to display additional match details and a delete button. The MatchAdministratorResponse model was created to carry the necessary data from the server to the client."