Add delete match functionality and match details page navigation
Added the delete match functionality for the Administrator role in the MatchController. The match, Elo Histories, and associated players' Elo are deleted if a valid matchId is passed. Also made an adjustment in the MatchesManager.razor view, separating the match content into a clickable div to navigate to a detailed match page and added a delete button for deleting the match.