"Add match simulation functionality and optimize Elo calculations"
This update includes the addition of a match simulation feature under the Matches Manager and associated button in the MatchesManager.razor file. This simulation feature is built in the MatchController.cs file with calculations for Elo ratings change. Changes are also made to optimize the calculation of Elo ratings after a match in the CalculateEloResultAfterMatchCommandHandler.cs.

To enhance code readability and maintenance, unnecessary printing lines and comments have been removed from the MatchDetail.razor and MatchController.cs files. A minor user interface improvement is also added, changing the greeting from <h1> to <h2> in Index.razor.

These changes aim to improve the user experience by offering a simulated match process along with a more efficient and clean codebase.