Rename EloHistory column names
In the EloHistory table, the names of two columns were updated: 'PreviousElo' to 'OldElo', and 'CurrentElo' to 'NewElo'. This change was made to improve clarity when referring to changes in player Elo across the system. Consequently, all affected codes including data migrations, controllers, and areas were updated to reflect this change.