Add real-time updates to match creation
Introduced the implementation of SignalR to add real-time updates to match creation. This involves configuring SignalR in the server's startup configuration and integrating it in ‘CreateAMatch.razor’ and ‘AllMatches.razor’ on the client side. The user will now see the fresh list of matches as soon as a new match is created by any other user. The change was necessary to enhance the user experience and ensure users have the latest information. Also added the necessary dependency, 'Microsoft.AspNetCore.SignalR.Client', to the project.