Add SignalR functionality to MatchDetail page
The change introduces SignalR functionality to the MatchDetails razor page for real-time updates. A HubConnection is initiated using .WithUrl and .Build then started in the OnInitializedAsync method. Implemented `IAsyncDisposable` to dispose of the connection when no longer needed. Whenever a match is updated or deleted, a message is sent via SignalR to notify connected clients about the change. Removed some redundant lines in CreateAMatch.razor.