Introduce SignalR to manage real-time updates
A SignalR client, `hubConnection`, has been introduced into the `MatchDetailPlayers.razor` page to push real-time updates to the client whenever a player is added or removed from a match. Now, when a player is added or removed, a message is sent via the SignalR hub (`"SendMessage"`). This hub has been connected during initialization to the `/allmatcheshub` endpoint. The `DisposeAsync` method has also been overriden to disconnect from the hub when the component is disposed.