Add real-time match update functionality
Implemented a new Hub 'MatchHub' to manage real-time updates specifically for individual match states. This Hub allows clients to join a match specific group and provides a method to send a message to all group members. The message is triggered on two actions in the 'MatchDetailPlayers' component: 'AddPlayer' and 'RemovePlayer' actions indicating an update to the Match. The new Hub 'MatchHub' is registered in Startup.cs. This enhances user experience by providing real-time match updates.