Refactor code and remove unnecessary using statements
This commit includes both refactoring and clean-up changes in the codebase.

Firstly, the logic of Match retrieval has been moved from the MatchController into a separate GetMatchQueryHandler. This was done to improve code modularity and separation of concerns, in accordance with the CQRS pattern.

The removed using statements (Color = Blazorise.Color, EventsManager.Shared.Requests from several files) were no longer necessary, as the referenced namespaces were not being used in the files. This clean-up helps maintain a neat codebase.