"Add match block/unblock feature"
This commit adds a new feature that allows administrators to block and unblock a match. It does so by introducing a new button in the match's manager page, which calls a new endpoint added in the MatchManager Controller. This action was added as a way to give administrators more control over the matches and prevent any possible malicious activities. A new property ("IsBlocked") is also added in the "MatchAdministratorResponse.cs" to keep track of the match's status.