Add block status to match view, change blocking method to PUT
In the MatchesManager.razor view, added a display for whether a match is blocked or not. In addition, changed the HTTP method for blocking/unblocking a match from POST to PUT in both the view and the MatchManagerController.cs. This is done to more accurately reflect the idempotent nature of the operation.