Add EloHistories to InitialPlayerSkillCalibration handler
The EloHistories data was missing in the InitialPlayerSkillCalibrationCommandHandler. This caused that player's initial Elo history was not correctly set during initial skill calibration. Now, the Handler includes EloHistories data of the player from the context and removes the initial Elo history after setting the new Elo. This ensures the proper computation and storing of player's Elo score history.