Refactor match component to improve player display
The layout and the player display in each match have been refactored. Layout changes include rearranging the information to improve visual hierarchy. We also abstracted away the player listing in Matches Component to a separate method which checks if a slot for a player is filled - this way we can show 'Open slot' if the slot is still free. The refactoring has made the layout cleaner and provides users with necessary information at first glance.