Initialize _loading to true in OnInitializedAsync()
The private boolean variable _loading was initialized to false at the time of declaration. This change moves the initialization of _loading to true at the start of the OnInitializedAsync method. This is done to ensure that loading state gets set prior to the invocation of asynchronous operations like starting a hub connection or getting match details, thus providing a more accurate representation of the loading state for UI feedback.