Remove default Elo value and handle failed user registration due to email issues
A new migration is added to remove the default Elo value (1500) from the Player model. This change is a design decision to avoid having an arbitrary default Elo rating for players.

Also, exception handling in the user registration process has been improved to catch and properly handle failed user registrations due to issues with sending confirmation email. Now, if an exception is thrown during the process of sending the confirmation email, the already created user is deleted and an error message is displayed.

EmailService has been modified to convert the recipient email to uppercase to handle some systems that are case sensitive by the email domain.