Set maximum width for several elements and comment out redundant hub connection
In this commit, a maximum width of 641px was set for several elements in multiple razor files to improve element wrapping on smaller displays. Furthermore, hubConnectionMatch, a redundant SignalR hub connection was found in the code, which was commented out to avoid unnecessary resource consumption. This included methods and condition checks associated with the hub connection. The changes implied will reduce resource use, address potential concurrency issues with having multiple hub connections and improve UI on smaller screens.