Update condition for showing "Odds" in MatchDetailPlayers
This commit changes the logic for when the "Odds" section is shown in the MatchDetailPlayers component. Previously, the "Odds" would appear as long as there were any players in both teams. Now it will only appear if the opposing team has any players. This prevents incorrectly showing "Odds" when only one team is selected or present.