Update level display in MatchDetail page
In the MatchDetail page, updated the level display to reflect color changes based on the user's eligibility for a match. A new method 'GetMinimumLevelColor' was added to determine the color. This change makes it easier to visually distinguish if the user qualifies for the match based on their ELO. A bug in GetMatchQueryHandler was also fixed where 'requesterElo' was not set correctly due to a null check on 'userId'. The color will now correctly reflect red for ineligible users and green for eligible ones.