Fix level comparison operator in match views
The comparison operator for displaying the minimum level in match details and in the all matches view was incorrect. It was previously using a less than ("<") sign when it should have used a greater than (">") sign. The "<" sign was misleading as it looked as if only users of lower levels could join the matches. This fix should provide clearer information about match requirements.