Add functionality to invite players to a Match
Added a new feature that enables the Match creator to invite players to their Match. Adjustments were made to the MatchDetailPlayers component to render an 'Invite' button when the requester is the Match creator. When the 'Invite' button is clicked, an invitation is sent to a player. A new HTTP PostAsync method was added to invite players. Changes were also made to the PlayerDto to include the player's Id, which is necessary for the invitation process. Furthermore, informative feedback was provided for user interactions with ToastService messages.