Add confirmation attribute to match invitations
A "Confirmed" field is added to the MatchPlayer model. This field holds a boolean value which results in a new migration being created. It allows creators of matches to invite players to join. A player becomes part of a match once they confirm the invitation. The MatchPlayer entity is updated accordingly and verified when confirming match scores. All affected files have been adjusted to accommodate this new attribute. The updates ensure the integrity of match scores by confirming the participating players.