Add match notifications for invited users
Added a new feature where users will be notified when they are invited to join a match in the Events Management App. Made changes in MatchController.cs and NotificationController.cs to handle match invitations respectively. Also added a new InvitedMatchesResponse.cs for sending response to front-end. Updated Index.razor to show notifications to invited users, and created a new component InvitedMatchesNotification.razor to display match invitations and corresponding actions. This improvement aims to boost user engagement by ensuring users are aware of all the matches they are invited to, and can accept the invitation straight away.