Changed invitation handling in game matches
Revised the existing match invitation system to allow users to accept or decline invitations. Previously, the system automatically confirmed every match after a user got invited and there was no option for the user to decline. Now, the function has been changed to a POST request that allows users to either accept or decline the match. In addition, the notification interface on the client side has been updated to show 'Accept' and 'Decline' buttons for each invitation. These modifications make the invitation process more interactive and user-controlled, leading to a better gaming experience.