Refactor player invitation search for match-specific functionality
The change refactors the route and functionality for searching players to invite to a given match. Previously, when filtering players, the method only used a text term to search by nickname. However, this led to matches in which players were already participating being suggested for invitation.

The new implementation takes a matchId parameter and discards players already engaged in that match from the query results. The front-end call has been adjusted accordingly.

Furthermore, to make the response more specific and decoupled, the `PlayerToInviteResponse` model was moved from the `PlayerController` to the shared responses directory.