Refactor RespondInvitation method logic
This commit refactors the logic in the RespondInvitation method in InvitedMatchesNotification.razor. Prior to this commit, acceptance or declination of the invitation request was initiated without clear user confirmation. Now, for declining invitations, user confirmation is required before the logic can proceed. For accepting invitations, confirmation is not needed which enhances the UX. Further, separate method 'CallInviteEndpoint' was created to avoid code duplication.