Add match invitations notifications feature
A new feature has been added to the match detail view. Now, users will be notified with an invitation from the match creator. They can choose to accept or decline the invitation. The new feature was introduced by adding related functionalities in both the client and the server. In the client-side, 'InvitedMatchNotification' component was created for inviting notifications and was used in 'MatchDetail' view. On the server-side, an 'IHaveOpenInvitation' property was added to the 'GetMatchQueryHandler' to track if a user has an open invitation for a match. This will aid in enhancing user interaction with our app by making match participation more engaging and comprehensive.