Prevent double join button press
Added a "_joining" boolean to disable the "Join" button after it's clicked in order to prevent double clicking and potential duplication of players in the teams. This is particularly useful in a situation where the join event takes significant time, stopping the users from making the redundant mistake of clicking join multiple times.