Change datatype of "Location" field to string in "Tournament" entity
The datatype of "Location" field in the "Tournament" entity was initially set as an integer, but for our application, it makes more sense to store it as a string since a location usually would be a written address or name. The datatype has been changed in the migration and type configuration. A default value "None" has also been set. Lastly, a new Razor component for viewing