Remove unused modal and add max teams to tournament
Deleted the PlayerDetailModal.razor component as it was not being used, improving codebase clarity. Also added a property 'MaxTeams' to the Tournament model to limit the number of teams that can participate in a tournament. This was necessary to ensure tournaments don't become overly crowded and unmanageable. Corresponding migration was created for the applied model changes. A 'Max Teams' field was also added to the Tournament request.