Add MaxTeams enum and incorporate into Tournament creation
This commit introduces a MaxTeams enum containing "Eight" and "Sixteen" as options, serving to limit the number of teams in a tournament. This provides more controlled and clear configuration options for tournament creation. The enum was incorporated in the TournamentRequest and TournamentController classes. Additionally, MaxTeams option was added to the CreateTournamentModal.ornament in the client side for user selection. A method in the Tournaments.razor file was also removed as it was not implemented and unnecessary.