Add functionality to track Tournament creator
Modified several files to add `Creator` property to `Tournament` model which is used to track and display creator of a particular tournament.
These changes are necessary to allow the feature of recognizing and displaying the creator of a tournament. Applied these updates in the existing Tournament model, Match model, Server database, and client side UI component.