Add feature to sign in to tournaments as a team
Added the ability for players to form teams and sign up together for tournaments. This has been implemented in the Tournament API and UI layers. Specifically, new CoupleResponse and TournamentSignInRequest classes were created; the former holds information about a team, and the latter is used to initiate the sign-up process. Changes were also made to the TournamentDetailResponse class and the TournamentMainTab frontend component to accomodate the team-based structure. The TournamentController class and the SignInTournamentModal frontend component saw significant changes as a result. Overall, these changes introduce more options for users participating in tournaments and help to streamline the sign-up process.