Add 'Delete Team' functionality in tournaments
A new functionality has been added to allow the tournament creator to delete teams. Guid Id is added to the CoupleResponse for identification. Also, a new HttpDelete method is added in the TournamentController permitting the tournament creator to remove teams. On the front-end side, a 'Delete Team' button is added to the TournamentMainTab component. If the current user is the tournament creator, they will see this button next to each team, clicking on which will call the newly added `DeleteTeam` method.