Add 'Price' column to 'Tournament' table
A migration is added to include a 'Price' column in the 'Tournament' table in the database. This is done to store the price information related to each tournament. Changes are also made in the relevant sections of the codebase to accommodate this update. This will allow event organizers to specify a price for participation in their tournaments, providing them with a means to cover expenses or collect funds for prizes.