Feat: Add court number and competitiveness to the match functionality
Added "IsCompetitive" and "CourtNumber" columns to the Match table for the purpose of distinguishing whether matches are competitive and specifying the court number on which the match will take place, respectively. Updated application data context and migrations to reflect these changes. Also adjusted frontend component EditMatch to include a selection for court number when match location is PadelBaseSalzburg.